Started working on the library. - Refactored the code a little. I moved almost all functions from the /bin/client.rs into /client/mod.rs. - I created a new file, /client/api.rs, that will contain the API for our client side. - In the /client/api.rs, I added a PiercerConfig struct, and a Piercer struct (empty for now), to be able to implement the first method for the library, which is **pub fn connect** -> that will be called as **Piercer::connect**. - In the **connect** method, I pretty much just copied the code from /client/bin.rs, from the **main** function. This is the first refactor attempt.
