+++ src/lib.rs
@@ -1,2 +1,4 @@
 pub mod client;
 pub mod signaling;
+
+pub use client::api::{Piercer, PiercerConfig};
